// src/invetory/interfaces/general-inventory.interface.ts
// Generated by https://quicktype.io

export interface GeneralInventoryDto {
    id:           string;
    beforeChange: BeforeChange;
}

export interface BeforeChange {
    id:          string;
    status:      string;
    observation: string;
    kilometers:  string;
    createdAt:   string;
    vehicle:     Vehicle;
}

export interface Vehicle {
    id:      string;
    model:   string;
    brand:   string;
    color:   string;
    chassis: string;
    images:  Image[];
}

export interface Image {
    id:  number;
    url: string;
}