// src/invetory/dto/update-general-inventory.dto.ts
// Generated by https://quicktype.io

export interface UpdateGeneralInventory {
  inventory: Inventory;
  inventory_checklist_items: { [key: string]: string };
}

export interface Inventory {
    status:      string;
    observation: string;
    kilometers:  number;
    vehicle:     Vehicle;
}

export interface Vehicle {
    model:   string;
    brand:   string;
    color:   string;
    chassis: string;
    images?:  any[];
}